UserStocks rebuild

+++ src/main/java/stockexchange/com/stockexchange/repository/UserRepository.java
@@ -5,8 +5,9 @@ import org.springframework.stereotype.Repository;
 import stockexchange.com.stockexchange.model.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long>{
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByUsername(String login);
-    User findById(Long id);
+
+    User findById(Integer id);
     User findByUsernameIgnoreCase(String username);
 }
